apps/backend/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.2",
    "fastapi>=0.128.0",
    "greenlet>=3.3.1",
    "gunicorn>=24.1.1",
    "psycopg[binary]>=3.3.2",
    "pydantic-settings>=2.11.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.46",
    "structlog>=25.5.0",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "python-dotenv>=1.2.1",
]
